renamed auth/token.js to auth/auth.js for consistency
